Repair Service Worker
Improved the updating method of the website, from the original interval checking for updates (which consumes a lot of network resources) to only checking the service worker when loading the interface, and when there is a new version, it will be updated automatically, without requiring users to update manually. Enhance the website experience.